fix: 优化性能

+++ src/system/node/background_color.rs
@@ -92,7 +92,7 @@ pub fn calc_background(
 
     let mut init_spawn_drawobj = Vec::new();
     for (node, background_color, layout, mut draw_list, background_color_change, layout_change) in query.p0().iter_mut() {
-        match draw_list.get(**render_type) {
+        match draw_list.get(**render_type as u32) {
             // background_color已经存在一个对应的DrawObj， 则修改color group
             Some(r) => {
                 let (mut draw_state, mut old_unit_quad, mut pipeline_meta) = match query_draw.get_mut(*r) {
@@ -177,7 +177,7 @@ pub fn calc_background(
 
 
                 // 建立Node对DrawObj的索引
-                draw_list.insert(**render_type, new_draw_obj);
+                draw_list.insert(**render_type as u32, new_draw_obj);
 
                 init_spawn_drawobj.push((new_draw_obj, draw_bundle));
             }
