+++ src/system/node/border_color.rs
@@ -86,7 +86,7 @@ pub fn calc_border_color(
 
     let mut init_spawn_drawobj = Vec::new();
     for (node_id, border_color, radius, layout, mut draw_list, background_color_change, radius_change, layout_change) in query.p0().iter_mut() {
-        match draw_list.get(**render_type) {
+        match draw_list.get(**render_type as u32) {
             // background_color已经存在一个对应的DrawObj， 则修改color group
             Some(r) => {
                 let (mut draw_state, mut pipeline_meta) = match query_draw.get_mut(*r) {
@@ -159,7 +159,7 @@ pub fn calc_border_color(
                     },
                 ));
                 // 建立Node对DrawObj的索引
-                draw_list.insert(**render_type, new_draw_obj);
+                draw_list.insert(**render_type as u32, new_draw_obj);
                 // draw_state_commands.insert(new_draw_obj, draw_state);
                 // fs_defines_commands.insert(new_draw_obj, fs_defines);
                 // is_unit_quad_commands.insert(new_draw_obj, BoxType::Border);
